Make the white balance picker more robust.
Earlier the change to neutral white balance could go out of sync
with when we wanted to sample, so we could come too early or too
late and get an already-corrected value.

This is also helped by the previous theme fixes for storing the
neutral_colors array as an upvalue in prepare_chain().