Make channel labels ignore their horizontal size policy.
This fixes an issue where the aspect ratio of the videos would be
all messed up due to the labels dictating the size. This could even
keep the window from being maximized if there were too many long
labels.

The best here would probably be some sort of marquee, but this is
an OK fix for now.