Redo x264 speed control presets without B-frame settings.
Seemingly, these can not be changed runtime, which kills our strategy
entirely for the lowest presets. Thus, these (unfortunately!) have to go;
they had no actual effect.

One _can_ set B-frame bias to -90, which gives much of the same effect,
but this is only picked up during reconfig if we have --sync-lookahead 0,
which seems to be much slower, so that won't really help either.