Support x264 >= 153.
Newer versions of x264 support multiple bit depths in one library,
so we don't need to load dynamically anymore. (On the other hand,
there's a new parameter we need to set.) Keep the old code around
for the time being, though, since we'd like to keep supporting
stretch for now.